+++ packages/components/src/DataList/DataList.css.d.ts
@@ -3,8 +3,13 @@ declare const styles: {
   readonly "titleContainer": string;
   readonly "headerFilters": string;
   readonly "headerTitles": string;
+  readonly "headerBatchSelect": string;
   readonly "listItem": string;
+  readonly "active": string;
+  readonly "listItemClickable": string;
   readonly "selectable": string;
+  readonly "selectAllCheckbox": string;
+  readonly "visible": string;
   readonly "selected": string;
   readonly "filtering": string;
   readonly "filteringSpinner": string;