+++ packages/components/src/DataList/DataList.types.ts
@@ -1,5 +1,6 @@
 import { ReactElement, ReactNode } from "react";
 import { IconNames } from "@jobber/design";
+import { XOR } from "ts-xor";
 import { Breakpoints } from "./DataList.const";
 import { ButtonProps } from "../Button";
 
@@ -47,7 +48,7 @@ export type DataListHeader<T extends DataListObject> = {
 
 export interface DataListSorting {
   readonly key: string;
-  readonly direction: "asc" | "desc";
+  readonly order: "asc" | "desc";
 }
 
 export interface DataListProps<T extends DataListObject> {
@@ -124,6 +125,8 @@ export interface DataListProps<T extends DataListObject> {
    * The list of Selected Item ids
    */
   readonly selected?: T["id"][];
+
+  readonly onSelectAll?: () => void;
 }
 
 export interface DataListLayoutProps<T extends DataListObject> {
@@ -182,15 +185,59 @@ export interface DataListContextProps<T extends DataListObject>
   readonly itemActionComponent?: ReactElement<DataListItemActionsProps<T>>;
 }
 
+export interface DataListLayoutContextProps {
+  readonly isInLayoutProvider: boolean;
+
+  /**
+   * Determine if the consumer of the DataList manually added an action to the
+   * layout. This is used to determine if the DataList should render the
+   * action column.
+   */
+  readonly hasInLayoutActions: boolean;
+  readonly setHasInLayoutActions: (state: boolean) => void;
+}
+
+export interface DataListLayoutActionsContextProps<T extends DataListObject> {
+  readonly activeItem?: T;
+}
+
 type Fragment<T> = T | T[];
 
-export interface DataListItemActionsProps<T extends DataListObject> {
+interface BaseDataListItemActionsProps<T extends DataListObject> {
   /**
    * The actions to render for each item in the DataList. This only accepts the
    * DataList.Action component.
    */
   readonly children?: Fragment<ReactElement<DataListActionProps<T>>>;
+
+  /**
+   * Callback when an item is clicked.
+   */
+  readonly onClick?: (item: T) => void;
 }
+
+interface DataListItemActionsPropsWithURL<T extends DataListObject>
+  extends BaseDataListItemActionsProps<T> {
+  /**
+   * If a normal page navigation is needed, use this prop to change the element
+   * to an `a` tag with an `href`.
+   */
+  readonly url?: string | ((item: T) => string);
+}
+
+interface DataListItemActionsPropsWithTo<T extends DataListObject>
+  extends BaseDataListItemActionsProps<T> {
+  /**
+   * If a React Navigation is needed, use this prop to use the `Link` component
+   * that comes with React Router.
+   */
+  readonly to?: string | ((item: T) => string);
+}
+
+export type DataListItemActionsProps<T extends DataListObject> = XOR<
+  DataListItemActionsPropsWithURL<T>,
+  DataListItemActionsPropsWithTo<T>
+>;
 
 export interface DataListActionProps<T extends DataListObject> {
   /**