+++ packages/components/src/DataList/components/DataListHeaderTile/SortingArrows.css.d.ts
@@ -1,6 +1,9 @@
 declare const styles: {
   readonly "arrowUp": string;
   readonly "arrowDown": string;
+  readonly "sortIcon": string;
+  readonly "active": string;
+  readonly "inactive": string;
 };
 export = styles;